requirements.txt
flask
torch
torchvision
Pillow
numpy
opencv-python
matplotlib
pytorch-grad-cam
gunicorn
fpdf
